Put the transaction manager in engine
