+++ pkg/vm/engine/disttae/engine.go
@@ -18,14 +18,11 @@ import (
 	"context"
 	"time"
 
+	"github.com/matrixorigin/matrixone/pkg/common/moerr"
 	"github.com/matrixorigin/matrixone/pkg/pb/logservice"
 	"github.com/matrixorigin/matrixone/pkg/txn/client"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine"
 )
-
-type Engine struct {
-	getClusterDetails GetClusterDetailsFunc
-}
 
 type GetClusterDetailsFunc = func() (logservice.ClusterDetails, error)
 
@@ -35,6 +32,7 @@ func New(
 ) *Engine {
 	return &Engine{
 		getClusterDetails: getClusterDetails,
+		txns:              make(map[string]*Transaction),
 	}
 }
 
@@ -60,6 +58,33 @@ func (e *Engine) Delete(context.Context, string, client.TxnOperator) error {
 	panic("unimplemented")
 }
 
+// hasConflict used to detect if a transaction on a cn is in conflict,
+// currently an empty implementation, assuming all transactions on a cn are conflict free
+func (e *Engine) HasConflict(txn *Transaction) bool {
+	return false
+}
+
+func (e *Engine) PreCommit(ctx context.Context, op client.TxnOperator) error {
+	txn := e.getTransaction(op)
+	if txn == nil {
+		return moerr.New(moerr.ErrTxnClosed, "the transaction has been committed or aborted")
+	}
+	defer e.delTransaction(txn)
+	if e.HasConflict(txn) {
+		return moerr.New(moerr.ErrTxnWriteConflict, "write conflict")
+	}
+	return nil
+}
+
+func (e *Engine) Rollback(ctx context.Context, op client.TxnOperator) error {
+	txn := e.getTransaction(op)
+	if txn == nil {
+		return moerr.New(moerr.ErrTxnClosed, "the transaction has been committed or aborted")
+	}
+	defer e.delTransaction(txn)
+	return nil
+}
+
 func (e *Engine) Nodes() (engine.Nodes, error) {
 	clusterDetails, err := e.getClusterDetails()
 	if err != nil {
@@ -81,4 +106,33 @@ func (e *Engine) Nodes() (engine.Nodes, error) {
 func (e *Engine) Hints() (h engine.Hints) {
 	h.CommitOrRollbackTimeout = time.Minute * 5
 	return
+}
+
+func (e *Engine) newTransaction(op client.TxnOperator) {
+	txn := &Transaction{
+		readOnly: false,
+		meta:     op.Txn(),
+		operator: op,
+	}
+	e.addTransaction(txn)
+}
+
+func (e *Engine) getTransaction(op client.TxnOperator) *Transaction {
+	e.Lock()
+	defer e.Unlock()
+	return e.txns[string(op.Txn().ID)]
+}
+
+func (e *Engine) addTransaction(txn *Transaction) {
+	e.Lock()
+	defer e.Unlock()
+	if _, ok := e.txns[string(txn.meta.ID)]; !ok {
+		e.txns[string(txn.meta.ID)] = txn
+	}
+}
+
+func (e *Engine) delTransaction(txn *Transaction) {
+	e.Lock()
+	defer e.Unlock()
+	delete(e.txns, string(txn.meta.ID))
 }
